Add files via upload
Learning and improving the Security Operation's capabilities using Threat Intelligence. 
Knowing your current needs and capabilities is vital to maximising the usage of Threat Intelligence.
Implementation of Intelligence-driven prevention and detection.
Revisiting Sigma, Uncoder and Elastalert for the example tooling of intelligence-driven detection. 
Focused on leveraging Threat Intelligence to improve the overall capabilities of a Security Operations Center, ranging from detection capabilities to preventive mechanisms that mitigate entirely known threats.

In this world of ever-evolving threats, we need to utilise every ounce of knowledge to protect our organisation, including all the intelligence provided by organisations that serve as Producers and bits of meaningful feedback from Consumers. Through collaboration, we can protect and improve the security posture of every organisation.